Split into two projects with separate assemblies to make initialization work for both Commerce and Cms
