+++ Optimizely.TestContainers.Commerce.Tests/Models/Commerce/TestProduct.cs
@@ -1,8 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using EPiServer.Commerce.Catalog.ContentTypes;
 using EPiServer.Commerce.Catalog.DataAnnotations;
+using EPiServer.Core;
+using EPiServer.DataAbstraction;
+using EPiServer.DataAnnotations;
 
-namespace Optimizely.TestContainers.Models.Commerce;
+namespace Optimizely.TestContainers.Commerce.Tests.Models.Commerce;
 
 [CatalogContentType(
     GUID = "0B06DE9B-6AE3-40FB-909E-E718CCC260AE",
