--- src/Web/StartupWithCms.cs
@@ -1,43 +1,0 @@
-using EPiServer.Cms.Shell;
-using EPiServer.Cms.UI.AspNetIdentity;
-using EPiServer.Scheduler;
-using EPiServer.Web.Routing;
-
-namespace Optimizely.TestContainers;
-
-public class StartupWithCms(IWebHostEnvironment webHostingEnvironment)
-{
-    public void ConfigureServices(IServiceCollection services)
-    {
-        if (webHostingEnvironment.IsDevelopment())
-        {
-            AppDomain.CurrentDomain.SetData("DataDirectory", Path.Combine(webHostingEnvironment.ContentRootPath, "App_Data"));
-
-            services.Configure<SchedulerOptions>(options => options.Enabled = false);
-        }
-
-        services
-            .AddCmsAspNetIdentity<ApplicationUser>()
-            .AddCms()
-            .AddAdminUserRegistration()
-            .AddEmbeddedLocalization<StartupWithCms>();
-    }
-
-    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-    {
-        if (env.IsDevelopment())
-        {
-            app.UseDeveloperExceptionPage();
-        }
-
-        app.UseStaticFiles();
-        app.UseRouting();
-        app.UseAuthentication();
-        app.UseAuthorization();
-
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.MapContent();
-        });
-    }
-}
