+++ test/OptimizelyTestContainers.Tests/OptimizelyTestContainers.Tests.csproj
@@ -35,6 +35,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\Optimizely.TestContainers.Shared\Optimizely.TestContainers.Shared.csproj" />
       <ProjectReference Include="..\..\src\Web\Optimizely.TestContainers.csproj" />
     </ItemGroup>
